第二次提交，现在没有出现error了，只是空间问题依旧没有解决，但是最主要还是时间复杂度的问题
#	状态	时间	内存
1	Accepted	0 ms	756 KB
2	Accepted	0 ms	792 KB
3	Accepted	0 ms	792 KB
4	Accepted	0 ms	784 KB
5	Accepted	4 ms	788 KB
6	Accepted	180 ms	1084 KB
7	Accepted	180 ms	1084 KB
8	Accepted	184 ms	1084 KB
9	Accepted	184 ms	1084 KB
10	Accepted	336 ms	1260 KB
11	Accepted	336 ms	1256 KB
12	Accepted	340 ms	1260 KB
13	Accepted	340 ms	1264 KB
14	Accepted	344 ms	1260 KB
15	Accepted	340 ms	1256 KB
16	Memory Limit Exceeded	-	-
17	Time Limit Exceeded	-	-
18	Time Limit Exceeded	-	-
19	Time Limit Exceeded	-	-
20	Time Limit Exceeded	-	-

+++ 03.c
@@ -5,8 +5,9 @@
 //对于学者城市，存一个数组，索引为城市编号，值为学者数量。
 //策略：去一个城市和0号城市谁更便宜就去哪里。
 //要注意的一些情况：1.1号城市单独考虑，2.起点终点同城市的情况
+//优化：1.先改改short
 int sum=-1,dst_city=-1;
-void optimize_city(int **dst,int m)
+void optimize_city(short **dst,int m)
 {
     for(int k=0;k<m;k++)
     {
@@ -31,7 +32,7 @@ void optimize_city(int **dst,int m)
         }
     }
 }
-int calc_sum(int **dst,int *s_city,int m,int city)
+int calc_sum(short **dst,int *s_city,int m,int city)
 {
     int temp_sum=0;
     for(int i=1;i<m;i++)//从1开始，0号的不需要走
@@ -59,10 +60,10 @@ int main()
 {
     int n,m,k;//参会者数量n、城市数量m、航班数量k，在程序中，a城市对应数组下标a-1
     scanf("%d%d%d",&n,&m,&k);
-    int **dst=(int **)malloc(m*sizeof(int *));
+    short **dst=(short **)malloc(m*sizeof(short *));
     for(int j=0;j<m;j++)
     {
-        dst[j]=(int *)calloc(m,sizeof(int));
+        dst[j]=(short *)calloc(m,sizeof(short));
     }
     for(int i=0;i<k;i++)
     {
